Updated command

+++ provision-lms.sh
@@ -17,8 +17,10 @@ DIR="$( cd "$( dirname "$( dirname "${BASH_SOURCE[0]}" )" )" && pwd )"
 
 #docker-compose exec lms bash -c 'rm -Rf /edx/app/edxapp/edx-platform/node_modules'
 docker-compose exec lms bash -c 'source /edx/app/edxapp/edxapp_env && cd /edx/app/edxapp/edx-platform && NO_PYTHON_UNINSTALL=1 paver install_prereqs'
-rm -Rf ${DIR}/edx-platform/node_modules
-docker-compose exec lms bash -c 'source /edx/app/edxapp/edxapp_env && cd /edx/app/edxapp/edx-platform && NO_PYTHON_UNINSTALL=1 paver install_prereqs'
+if [ $? -ne 0 ]; then
+    rm -Rf ${DIR}/edx-platform/node_modules
+    docker-compose exec lms bash -c 'source /edx/app/edxapp/edxapp_env && cd /edx/app/edxapp/edx-platform && NO_PYTHON_UNINSTALL=1 paver install_prereqs'
+fi
 
 #Installing prereqs crashes the process
 docker-compose restart lms
